requirements.txt
# Google APIs
gspread>=5.12.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.111.0  # For Gmail

# Claude API
anthropic>=0.18.0

# Utilities
python-dotenv>=1.0.0

# Development (optional)
pytest>=7.4.0
